fix: 구독 목록이 뜨지 않는 오류 수정

+++ src/main/java/com/hy/photogram/web/dto/subscribe/SubscribeDto.java
@@ -16,6 +16,6 @@ public class SubscribeDto {
     private BigInteger id;
     private String username;
     private String profile_image_url;
-    private Integer subscribe_state;
-    private Integer equal_user_state;
+    private BigInteger subscribe_state;
+    private BigInteger equal_user_state;
 }
